# autoassure-server-sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "A2.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/applications/{applicationId}/runs": {
      "post": {
        "tags": [
          "Runs"
        ],
        "operationId": "CreateRun",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "400": {
            "description": "EnvironmentId does not reference an Environment belonging to this\nApplication, ScenarioIds contains a duplicate, or ScenarioIds contains an\nid that does not\nreference a Scenario belonging to this Application. Returns 400 when the request fails a validation constraint.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Application with the given applicationId exists in the caller's\nOrganization, or\nit no longer exists (deleted after this request started).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      },
      "get": {
        "tags": [
          "Runs"
        ],
        "operationId": "ListRuns",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/applications/{applicationId}/runs/running": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Lists the Runs currently Running for this Application, strongly consistent\n-- a Run that\njust started is never briefly missing from this result, unlike\nTask&lt;ActionResult&lt;IReadOnlyList&lt;RunSummaryResponse&gt;&gt;&gt; RunsController.List(Guid applicationId).",
        "operationId": "ListRunningRuns",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunningRunResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunningRunResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunningRunResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/applications/{applicationId}/runs/{runId}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "operationId": "GetRunById",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Run with the given runId exists in this Application, in the caller's\nOrganization.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{applicationId}/runs/{runId}/start": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Starts a Run and returns it with Environment variable values unmasked --\nthe last\nchance to see them unmasked. After the Run starts, every response masks\nthem, for\nsecurity.",
        "operationId": "StartRun",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Run with the given runId exists in this Application, in the caller's\nOrganization.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The Run's Status is not Pending.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      }
    },
    "/applications/{applicationId}/runs/{runId}/end": {
      "post": {
        "tags": [
          "Runs"
        ],
        "operationId": "EndRun",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndRunRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EndRunRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EndRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "TerminalStatus is Pending or Running. Returns 400 when the request fails a validation constraint.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Run with the given runId exists in this Application, in the caller's\nOrganization.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The Run's Status is not Running.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      }
    },
    "/applications/{applicationId}/runs/{runId}/heartbeat": {
      "post": {
        "tags": [
          "Runs"
        ],
        "operationId": "UpdateRunHeartbeat",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "No Run with the given runId exists in this Application, in the caller's\nOrganization.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The Run's Status is not Running.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "required": [
          "scenarioIds",
          "environmentId"
        ],
        "type": "object",
        "properties": {
          "scenarioIds": {
            "maxItems": 96,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "environmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "description": "Request body to start a Manual Run of one or more Scenarios against an\nEnvironment. Every id\nin ScenarioIds must reference a Scenario belonging to the Application named\nin the URL."
      },
...
      "RunResponse": {
        "required": [
          "id",
          "applicationId",
          "trigger",
          "status",
          "totalActivityCount",
          "passedActivityCount",
          "failedActivityCount",
          "skippedActivityCount",
          "environment",
          "scenarios",
          "lastSeq",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "trigger": {
            "$ref": "#/components/schemas/RunTrigger"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "totalActivityCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "passedActivityCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "failedActivityCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "skippedActivityCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "environment": {
            "$ref": "#/components/schemas/RunEnvironmentSnapshotResponse"
          },
          "scenarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunScenarioSnapshotResponse"
            },
            "description": "One snapshot per Scenario the Run ran, in no particular order."
          },
          "lastSeq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The highest sequence number appended to this Run's status update log so\nfar. A client\nthat already holds up to this sequence has nothing new to poll for.",
            "format": "int64"
          },
          "triggeredByUserId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Who triggered this Run. Null for a Scheduled Run -- a system timer has\nno user id.",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastHeartbeatAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When the owning worker last proved it was alive. Null while Status is\nPending. A Run\nstuck on Running with an old LastHeartbeatAt has likely lost its worker --\nStatus alone does not\ntell you that.",
            "format": "date-time"
          }
        },
        "description": "A Run's identity, execution state and what it ran, as returned to the\nclient. Never carries\nthe status update log -- LastSeq and Status are what tell a client whether\nit is worth polling List\nRun Status Updates and when to stop. ApplicationId is included even under\nthe nested\n`/applications/{applicationId}/runs/{runId}` route because the\nauthoring create route\n(`POST /scenarios/{scenarioId}/runs`) is flat and returns this same\nshape -- without it, a client\nfollowing an authoring Run would have no way to build its polling URLs."
      },
...
      "RunTrigger": {
        "type": "integer",
        "description": "Where a Run came from, as returned to the client. Affects retention and\nwhether the Run\nshows up in the Application's Runs panel -- nothing about how it executes."
      },
...
    }
  },
...
}